project/requirements.txt
#Base
pycodestyle==2.9.1
psycopg2==2.8.6,<2.9
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
asgiref==3.5.2
autopep8==1.7.0
djangorestframework-simplejwt==5.2.1

# Django
django==3.2.3
django-cors-headers==3.13.0
django-filter==22.1

# Django REST Framework
djangorestframework==3.13.1